+++ metadata.rb
@@ -7,7 +7,7 @@ maintainer_email 'j@p4nt5.com'
 license          'apache2'
 description      'Installs Spotify'
 long_description 'Installs Spotify'
-version          '0.1.0'
+version          '0.1.1'
 
 depends          'dmg', '~> 2.2'
 depends          'windows', '~> 1.37'