+++ spec/support/cookbooks/dmg/metadata.rb
@@ -1,9 +1,0 @@
-# Encoding: UTF-8
-
-name             'dmg'
-maintainer       'test'
-maintainer_email 'example@example.com'
-license          'apache2'
-description      'dmg'
-long_description 'dmg'
-version          '0.0.1'