+++ spec/support/cookbooks/resource_spotify_app_test/metadata.rb
@@ -1,0 +1,16 @@
+# Encoding: UTF-8
+
+name             'resource_spotify_app_test'
+maintainer       'Jonathan Hartman'
+maintainer_email 'j@p4nt5.com'
+license          'apache2'
+description      'spotify_app resource test wrapper'
+long_description 'spotify_app resource test wrapper'
+version          '0.0.1'
+
+depends          'spotify'
+
+supports         'mac_os_x'
+supports         'windows'
+supports         'ubuntu'
+supports         'debian'