+++ spec/support/cookbooks/resource_spotify_app_test/recipes/default.rb
@@ -1,0 +1,3 @@
+# Encoding: UTF-8
+
+spotify_app 'default'