+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 Spotify Cookbook CHANGELOG
 ==========================
 
+v?.?.? (????-??-??)
+-------------------
+
 v0.1.0 (2015-05-26)
 -------------------
 - Initial release! Supports OS X, Windows, and Ubuntu/Debian!