+++ CHANGELOG.md
@@ -7,6 +7,7 @@ Unreleased
 - Update to new repo key used for Ubuntu/Debian installs
 - Bump apt dependency from 2.x to 3.x
 - Forcibly install the xdg-utils dependency on Debian platforms
+- Convert everything to Chef 12.5 custom resources
 
 v0.1.0 (2015-05-26)
 -------------------